// ui/saltcorn-ui/vendor/saltcorn-data/base-plugin/types.ts
const show_with_html = {
  configFields: [
    {
      input_type: "code",
      name: "code",
      label: "HTML",
      sublabel: "Access the value with <code>{{ it }}</code>.",
      default: "",
      attributes: { mode: "text/html" },
    },
  ],
  isEdit: false,
  description: "Show value with any HTML code",
  run: (v: any, req: Req, attrs: any = {}) => {
    const ctx: any = { ...getState()!.eval_context };
    ctx.it = v;
    const rendered = interpolate(
      attrs?.code,
      ctx,
      req?.user,
      "show_with_html code"
    );
    return rendered;
  },
};
// ...
const to_locale_string = {
  description: "Show as in locale-sensitive representation",
  configFields: (field: any) => [
    {
      type: "String",
      name: "locale",
      label: "Locale",
      sublabel: "Blank for default user locale",
    },
    {
      type: "String",
      name: "style",
      label: "Style",
      required: true,
      attributes: {
        options: ["decimal", "currency", "percent", "unit"],
      },
    },
    {
      type: "Integer",
      name: "minimumFractionDigits",
      label: "Min Fraction Digits",
      sublabel: "Pad decimals to this many places (e.g. 2 → always show cents)",
      attributes: { min: 0, max: 20 },
    },
    {
      type: "Integer",
      name: "maximumFractionDigits",
      label: "Max Fraction Digits",
      attributes: {
        min: 0,
      },
    },
    {
      type: "Integer",
      name: "maximumSignificantDigits",
      label: "Max Significant Digits",
      attributes: {
        min: 0,
      },
    },
    {
      type: "String",
      name: "currency",
      label: "Currency",
      sublabel: "ISO 4217. Example: USD or EUR",
      required: true,
      showIf: { style: "currency" },
    },
    {
      type: "String",
      name: "currencyDisplay",
      label: "Currency display",
      required: true,
      showIf: { style: "currency" },
      attributes: {
        options: ["symbol", "code", "narrowSymbol", "name"],
      },
    },
    {
      type: "String",
      name: "currencySign",
      label: "Currency sign",
      sublabel: "Accounting: show negatives as (1,234.57) instead of -1,234.57",
      showIf: { style: "currency" },
      attributes: { options: ["standard", "accounting"] },
    },
    {
      type: "String",
      name: "signDisplay",
      label: "Sign display",
      sublabel:
        "always: show + on positives; exceptZero: show sign except on 0",
      attributes: {
        options: ["auto", "always", "exceptZero", "negative", "never"],
      },
    },
    {
      type: "String",
      name: "unit",
      label: "Unit",
      required: true,
      showIf: { style: "unit" },
      attributes: {
        options: [
          "acre",
          "bit",
          "byte",
          "celsius",
          "centimeter",
          "day",
          "degree",
          "fahrenheit",
          "fluid-ounce",
          "foot",
          "gallon",
          "gigabit",
          "gigabyte",
          "gram",
          "hectare",
          "hour",
          "inch",
          "kilobit",
          "kilobyte",
          "kilogram",
          "kilometer",
          "liter",
          "megabit",
          "megabyte",
          "meter",
          "microsecond",
          "mile",
          "mile-scandinavian",
          "milliliter",
          "millimeter",
          "millisecond",
          "minute",
          "month",
          "nanosecond",
          "ounce",
          "percent",
          "petabyte",
          "pound",
          "second",
          "stone",
          "terabit",
          "terabyte",
          "week",
          "yard",
          "year",
        ],
      },
    },
    {
      type: "String",
      name: "unitDisplay",
      label: "Unit display",
      required: true,
      showIf: { style: "unit" },
      attributes: {
        options: ["short", "narrow", "long"],
      },
    },
  ],
  isEdit: false,
  run: (v: any, req: Req, attrs: any = {}) => {
    const v1 = typeof v === "string" ? +v : v;
    if (typeof v1 === "number" && !isNaN(v1)) {
      const locale_ = attrs.locale || locale(req);
      return v1.toLocaleString(locale_, {
        style: attrs.style,
        currency: attrs.currency,
        currencyDisplay: attrs.currencyDisplay,
        currencySign: attrs.currencySign || undefined,
        unit: attrs.unit,
        unitDisplay: attrs.unitDisplay,
        signDisplay: attrs.signDisplay || undefined,
        minimumFractionDigits:
          attrs.minimumFractionDigits === 0
            ? 0
            : attrs.minimumFractionDigits || undefined,
        maximumSignificantDigits:
          attrs.maximumSignificantDigits === 0
            ? 0
            : attrs.maximumSignificantDigits || undefined,
        maximumFractionDigits:
          attrs.maximumFractionDigits == 0
            ? 0
            : attrs.maximumFractionDigits || undefined,
      });
    } else return "";
  },
};
// ...
const color = {
  /** @type {string} */
  name: "Color",
  description: "Colors, defined as Red, Green and Blue with 256 level each",
  /** @type {string} */
  sql_name: "text",
  js_type: "string",

  /**
   * @returns {function}
   */
  contract: () => is.str,
  /**
   * @namespace
   * @category saltcorn-data
   * @subcategory types / color
   */
  fieldviews: {
    /**
     * @namespace
     * @category saltcorn-data
     * @subcategory types / color
     */
    show: {
      isEdit: false,
      description: "A box filled with the color",
      run: (s: any) =>
        s
          ? div({
              class: "color-type-show",
              style: `background: ${s};`,
            })
          : "",
    },
    show_with_html,
    /**
     * @namespace
     * @category saltcorn-data
     * @subcategory types / color
     */
    edit: {
      isEdit: true,
      blockDisplay: true,
      description: "Simple color picker",
      run: (nm: any, v: any, attrs: any, cls: any, required: any, field: any) =>
        input({
          type: "color",
          class: ["form-control", cls],
          disabled: attrs.disabled,
          readonly: attrs.readonly,
          onChange: attrs.onChange,
          "data-fieldname": text_attr(field.name),
          name: text_attr(nm),
          id: `input${text_attr(nm)}`,
          ...(isdef(v) && { value: text_attr(v) }),
        }),
    },
  },
  /** @type {object[]} */
  attributes: [],
  /**
   * @param {object} v
   * @returns {object}
   */
  read: (v: any) => {
    switch (typeof v) {
      case "string":
        return v;
      default:
        return undefined;
    }
  },
  /**
   * @returns {boolean}
   */
  validate: () => (x: any) => {
    return true;
  },
};

/**
 * Float type
 * @namespace
 * @category saltcorn-data
 * @subcategory types / float
 */
const float = {
  /** @type {string} */
  name: "Float",
  description: "Floating-point numbers",
  /** @type {string} */
  sql_name: "double precision",
  js_type: "number",

  /**
   * @param {object} opts
   * @param {number} opts.min
   * @param {number} opts.max
   * @returns {function}
   */
  contract: ({ min, max }: any) => is.number({ lte: max, gte: min }),

  distance_operators: { near: sqlFun("ABS", sqlBinOp("-", "target", "field")) },

  /**
   * @namespace
   * @category saltcorn-data
   * @subcategory types / float
   */
  fieldviews: {
    /**
     * @namespace
     * @category saltcorn-data
     * @subcategory types / float
     */
    show: {
      isEdit: false,
      description: "Show number with no additional formatting",
      run: (s: any) => text(s),
    },
    /**
     * @namespace
     * @category saltcorn-data
     * @subcategory types / float
     */
    edit: {
      isEdit: true,
      blockDisplay: true,
      description: "Number input",
      configFields: [
        {
          name: "readonly",
          label: "Read-only",
          type: "Bool",
        },
      ],
      run: (nm: any, v: any, attrs: any, cls: any, required: any, field: any) =>
        input({
          type: "number",
          class: ["form-control", cls],
          name: text_attr(nm),
          "data-fieldname": text_attr(field.name),
          disabled: attrs.disabled,
          readonly: attrs.readonly,
          onChange: attrs.onChange,
          required: !!required,
          step: attrs.decimal_places
            ? Math.round(
                Math.pow(10, -attrs.decimal_places) *
                  Math.pow(10, attrs.decimal_places)
              ) / Math.pow(10, attrs.decimal_places)
            : "any",
          id: `input${text_attr(nm)}`,
          ...(isdef(attrs.max) && { max: attrs.max }),
          ...(isdef(attrs.min) && { min: attrs.min }),
          ...(isdef(v) && { value: text_attr(v) }),
        }),
    },
    number_slider: number_slider("Float"),
    range_interval: range_interval("Float"),
    progress_bar: progress_bar("Float"),
    heat_cell: heat_cell("Float"),
    above_input: float_number_limit("gte"),
    below_input: float_number_limit("lte"),
    to_locale_string,
    show_with_html,
  },
  /** @type {object[]} */
  attributes: [
    { name: "min", label: "Minimum", type: "Float", required: false },
    { name: "max", label: "Maximum", type: "Float", required: false },
    { name: "units", label: "Units", type: "String", required: false },
    {
      name: "decimal_places",
      label: "Decimal places",
      type: "Integer",
      required: false,
    },
  ],
  /**
   * @param {object} v
   * @returns {number|string|undefined}
   */
  read: (v: any) => {
    switch (typeof v) {
      case "number":
        return v;
      case "string":
        const stripped = v.replace(/[^0-9.\-e]+/g, "");
        if (!stripped) return undefined;
        const parsed = Number(stripped);
        return isNaN(parsed) ? undefined : parsed;
      default:
        return undefined;
    }
  },
  /**
   * @param {object} param
   * @returns {object|boolean}
   */
  validate:
    ({ min, max }: any) =>
    (x: any) => {
      if (isdef(min) && x < min) return { error: `Must be ${min} or higher` };
      if (isdef(max) && x > max) return { error: `Must be ${max} or less` };
      return true;
    },

  /**
   * check if two float values are equal within a given precision
   * @param a
   * @param b
   * @param opts
   * @returns true or false
   */
  equals: (a: any, b: any, { decimal_places }: any) => {
    return Math.abs(a - b) < Math.pow(10, -decimal_places) / 2;
  },
};